Bumped file dep to 6.0.0-nullsafety.1 to fix versioning errors in Flutter 1.22 (dev/master channel)

+++ flutter_cache_manager/lib/src/compat/file_service_compat.dart
@@ -79,5 +79,5 @@ class CompatFileServiceGetResponse implements FileServiceResponse {
   }
 
   @override
-  int get statusCode => legacyResponse.statusCode as int;
+  int get statusCode => legacyResponse.statusCode;
 }
